services/voice/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc-voice-agent"
version = "0.1.0"
description = "Voice agent (Scarlett) for the A.R.C. platform — LiveKit, STT, TTS, NATS bridge"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # LiveKit agent SDK
    "livekit-agents>=0.8",
    # Speech-to-Text
    "faster-whisper>=1.0",
    # Text-to-Speech (piper Python bindings; binary installed separately in Dockerfile)
    "piper-tts>=1.2",
    # Messaging
    "nats-py>=2.9",
    # pulsar-client==3.10.0: cp313 wheels verified for linux/amd64 and linux/arm64
    "pulsar-client==3.10.0",
    # Cache (Redis — session state)
    "redis>=5.0",
    # Observability
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27",
    "structlog>=24.4",
    # Resilience
    "tenacity>=9.0",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.11",
    "pyyaml>=6.0",
    "types-PyYAML",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

# pulsar-client has no type stubs — same pattern as reasoner service
[[tool.mypy.overrides]]
module = ["voice.pulsar_events"]
disable_error_code = ["union-attr"]

# FastAPI route decorators are untyped in the FastAPI stubs
[[tool.mypy.overrides]]
module = [
    "voice.stt_router",
    "voice.tts_router",
    "voice.health_router",
]
disable_error_code = ["untyped-decorator"]

